Moving exceptions file to more logically named file (#2407)
